+++ test/test_livingapi.py
@@ -2834,21 +2834,69 @@ def test_record_attachments_on_demand(handler, config_apps):
 def test_app_templates_on_demand(handler, config_persons):
 	if not isinstance(handler, PythonHTTP):
 		handler.make_internaltemplate(
-			identifier="test_livingapi_app_template_on_demand",
-			source="<?print app.t_test_app_template_on_demand_internal.name?>",
+			identifier="test_livingapi_app_template_on_demand_internal",
+			source="<?print app.t_test_livingapi_app_template_on_demand_internal.name?>",
 		)
 
 		vt = handler.make_viewtemplate(
 			identifier="test_livingapi_app_templates_on_demand",
 			source="""
-				<?render app.templates.test_app_shortcuts_internal(app=app)?>
-				<?render app.t_test_app_shortcuts_internal(app=app)?>
+				<?render app.templates.test_livingapi_app_template_on_demand_internal(app=app)?>
+				<?render app.t_test_livingapi_app_template_on_demand_internal(app=app)?>
 			"""
 		)
 
 		output = handler.renders(person_app_id(), template=vt.identifier)
 		expected = """
-			test_app_shortcuts_internal
-			test_app_shortcuts_internal
+			test_livingapi_app_template_on_demand_internal
+			test_livingapi_app_template_on_demand_internal
+		"""
+		assert lines(output) == lines(expected)
+
+
+def test_template_libraries(handler):
+	if not isinstance(handler, PythonHTTP):
+		vt = handler.make_viewtemplate(
+			identifier="template_libraries",
+			source="""
+				<?render globals.libs.la_static.templates.la_static_ul4()?>
+				<?render globals.libs.la_static.t_la_static_ul4()?>
+				<?render globals.l_la_static.templates.la_static_ul4()?>
+				<?render globals.l_la_static.t_la_static_ul4()?>
+			"""
+		)
+
+		output = handler.renders(person_app_id(), template=vt.identifier)
+		expected = """
+			<script src="/static/ul4/1.13.0/dist/umd/ul4.js"></script>
+			<script src="/static/ul4/1.13.0/dist/umd/ul4.js"></script>
+			<script src="/static/ul4/1.13.0/dist/umd/ul4.js"></script>
+			<script src="/static/ul4/1.13.0/dist/umd/ul4.js"></script>
+		"""
+		assert lines(output) == lines(expected)
+
+
+def test_chained_template_library(handler):
+	if not isinstance(handler, PythonHTTP):
+		vt = handler.make_viewtemplate(
+			identifier="chained_template_library",
+			source="""
+				<?render globals.app.cl_la_static.templates.la_static_ul4()?>
+				<?render globals.app.cl_la_static.t_la_static_ul4()?>
+				<?render globals.cl_la_static.templates.la_static_ul4()?>
+				<?render globals.cl_la_static.t_la_static_ul4()?>
+				<?print globals.app.cl_foo.identifier?>
+				<?print globals.cl_foo.identifier?>
+			"""
+		)
+
+		output = handler.renders(person_app_id(), template=vt.identifier)
+		expected = """
+			<script src="nosource"></script>
+			<script src="nosource"></script>
+			<script src="nosource"></script>
+			<script src="nosource"></script>
+			foo
+			foo
 		"""
 		assert lines(output) == lines(expected)