+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 HEAD (2022-07-07)
 -----------------
 
-*	Add support for template libraries.
+*	Add support for template libraries and template library chains.
 
 
 0.15 (2022-06-15)